Revert "Add Confluence Cloud connector (#66)"
This reverts commit 3a11fb4c44cbd9480f17fe5836483bc49d5dbcbe.
