+++ Gemfile
@@ -13,7 +13,6 @@ gem 'activesupport', '5.2.6'
 gem 'bson', '~> 4.2.2'
 gem 'mime-types', '= 3.1'
 gem 'tzinfo-data', '= 1.2022.1'
-gem 'nokogiri', '>= 1.13.4', :require => false
 
 group :test do
   gem 'rspec-collection_matchers', '~> 1.2.0'
@@ -39,7 +38,6 @@ gem 'sinatra-contrib', '~> 2.1.0'
 gem 'rack', '~> 2.2.3'
 gem 'forwardable', '~> 1.3.2'
 gem 'faraday', '~> 1.10.0'
-gem 'faraday_middleware', '= 1.0.0'
 gem 'httpclient', '~> 2.8.3'
 gem 'attr_extras', '~> 6.2.5'
 gem 'hashie', '~> 5.0.0'
