+++ lib/connectors_sdk/base/registry.rb
@@ -26,9 +26,7 @@ module ConnectorsSdk
 
     # loading plugins (might replace this with a directory scan and conventions on names)
     require_relative '../share_point/http_call_wrapper'
-    require_relative '../confluence_cloud//http_call_wrapper'
 
-    REGISTRY.register(ConnectorsSdk::SharePoint::HttpCallWrapper::SERVICE_TYPE, ConnectorsSdk::SharePoint::HttpCallWrapper)
-    REGISTRY.register(ConnectorsSdk::ConfluenceCloud::HttpCallWrapper::SERVICE_TYPE, ConnectorsSdk::ConfluenceCloud::HttpCallWrapper)
+    REGISTRY.register(ConnectorsSdk::SharePoint::SERVICE_TYPE, ConnectorsSdk::SharePoint::HttpCallWrapper)
   end
 end
