+++ lib/connectors_sdk/share_point/authorization.rb
@@ -6,12 +6,70 @@
 
 # frozen_string_literal: true
 
-require 'connectors_sdk/base/authorization'
+require 'connectors_shared'
+require 'signet'
+require 'signet/oauth_2'
+require 'signet/oauth_2/client'
 
 module ConnectorsSdk
   module SharePoint
-    class Authorization < ConnectorsSdk::Base::Authorization
+    class Authorization
       class << self
+        def authorization_url
+          'https://login.microsoftonline.com/common/oauth2/v2.0/authorize'
+        end
+
+        def token_credential_uri
+          'https://login.microsoftonline.com/common/oauth2/v2.0/token'
+        end
+
+        def authorization_uri(params)
+          missing = missing_fields(params, %w[client_id])
+          unless missing.blank?
+            raise ConnectorsShared::ClientError.new("Missing required fields: #{missing.join(', ')}")
+          end
+
+          params[:response_type] = 'code'
+          params[:additional_parameters] = { :prompt => 'consent' }
+          client = oauth_client(params)
+          client.authorization_uri.to_s
+        end
+
+        def access_token(params)
+          missing = missing_fields(params, %w[client_id client_secret code redirect_uri])
+          unless missing.blank?
+            raise ConnectorsShared::ClientError.new("Missing required fields: #{missing.join(', ')}")
+          end
+
+          params[:grant_type] = 'authorization_code'
+          client = oauth_client(params)
+          client.fetch_access_token
+        end
+
+        def refresh(params)
+          missing = missing_fields(params, %w[client_id client_secret refresh_token])
+          unless missing.blank?
+            raise ConnectorsShared::ClientError.new("Missing required fields: #{missing.join(', ')}")
+          end
+
+          params[:grant_type] = 'refresh_token'
+          client = oauth_client(params)
+          client.refresh!
+        rescue StandardError => e
+          ConnectorsShared::ExceptionTracking.log_exception(e)
+          raise ConnectorsShared::TokenRefreshFailedError
+        end
+
+        def oauth_client(params)
+          options = params.merge(
+            :authorization_uri => authorization_url,
+            :token_credential_uri => token_credential_uri,
+            :scope => oauth_scope
+          )
+          options[:state] = JSON.dump(options[:state]) if options[:state]
+          Signet::OAuth2::Client.new(options)
+        end
+
         def oauth_scope
           %w[
             User.ReadBasic.All
@@ -24,18 +82,8 @@ module ConnectorsSdk
           ]
         end
 
-        private
-
-        def authorization_url
-          'https://login.microsoftonline.com/common/oauth2/v2.0/authorize'
-        end
-
-        def token_credential_uri
-          'https://login.microsoftonline.com/common/oauth2/v2.0/token'
-        end
-
-        def additional_parameters
-          { :prompt => 'consent' }
+        def missing_fields(params, required = [])
+          Array.wrap(required).select { |field| params[field.to_sym].nil? }
         end
       end
     end
