+++ spec/connectors_sdk/sharepoint/authorization_spec.rb
@@ -6,18 +6,14 @@
 
 # frozen_string_literal: true
 
-describe ConnectorsSdk::Base::Authorization do
-  subject do
-    ConnectorsSdk::SharePoint::Authorization
-  end
-
+describe ConnectorsSdk::SharePoint::Authorization do
   describe '.authorization_uri' do
     let(:params) { { :client_id => 'client_id' } }
 
     context 'with invalid params' do
       %w[client_id].each do |param|
         it "raises ClientError for missing #{param}" do
-          expect { subject.authorization_uri(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
+          expect { described_class.authorization_uri(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
         end
       end
     end
@@ -27,7 +23,7 @@ describe ConnectorsSdk::Base::Authorization do
 
       it 'returns authorization uri' do
         allow_any_instance_of(Signet::OAuth2::Client).to receive(:authorization_uri).and_return(authorization_uri)
-        expect(subject.authorization_uri(params)).to eq(authorization_uri.to_s)
+        expect(described_class.authorization_uri(params)).to eq(authorization_uri.to_s)
       end
     end
   end
@@ -45,7 +41,7 @@ describe ConnectorsSdk::Base::Authorization do
     context 'with invalid params' do
       %w[client_id client_secret code redirect_uri].each do |param|
         it "raises ClientError for missing #{param}" do
-          expect { subject.access_token(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
+          expect { described_class.access_token(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
         end
       end
     end
@@ -60,7 +56,7 @@ describe ConnectorsSdk::Base::Authorization do
 
       it 'returns access token' do
         allow_any_instance_of(Signet::OAuth2::Client).to receive(:fetch_access_token).and_return(token_hash)
-        expect(subject.access_token(params)).to eq(token_hash)
+        expect(described_class.access_token(params)).to eq(token_hash)
       end
     end
   end
@@ -78,7 +74,7 @@ describe ConnectorsSdk::Base::Authorization do
     context 'with invalid params' do
       %w[client_id client_secret refresh_token].each do |param|
         it "raises ClientError for missing #{param}" do
-          expect { subject.access_token(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
+          expect { described_class.access_token(params.except(param.to_sym)) }.to raise_error(ConnectorsShared::ClientError)
         end
       end
     end
@@ -94,7 +90,7 @@ describe ConnectorsSdk::Base::Authorization do
 
         it 'returns access token' do
           allow_any_instance_of(Signet::OAuth2::Client).to receive(:refresh!).and_return(token_hash)
-          expect(subject.refresh(params)).to eq(token_hash)
+          expect(described_class.refresh(params)).to eq(token_hash)
         end
       end
 
@@ -102,7 +98,7 @@ describe ConnectorsSdk::Base::Authorization do
         let(:error) { 'error' }
         it 'returns authorization error' do
           allow_any_instance_of(Signet::OAuth2::Client).to receive(:refresh!).and_raise(Signet::AuthorizationError.new(error))
-          expect { subject.refresh(params) }.to raise_error(ConnectorsShared::TokenRefreshFailedError)
+          expect { described_class.refresh(params) }.to raise_error(ConnectorsShared::TokenRefreshFailedError)
         end
       end
     end
