+++ spec/connectors_sdk/sharepoint/http_call_wrapper_spec.rb
@@ -6,38 +6,67 @@
 
 # frozen_string_literal: true
 
+require 'active_support/core_ext/object/deep_dup'
 require 'connectors_sdk/share_point/http_call_wrapper'
-require 'connectors_sdk/base/http_call_wrapper'
-require 'connectors_sdk/office365/custom_client'
-require 'connectors_sdk/share_point/extractor'
+require 'connectors_sdk/base/custom_client'
+require 'connectors_sdk/base/adapter'
+require 'connectors_sdk/base/config'
+require 'connectors_sdk/base/extractor'
+require 'json'
+require 'time'
 
-describe ConnectorsSdk::Base::HttpCallWrapper do
-  let(:wrapper_class) { ConnectorsSdk::SharePoint::HttpCallWrapper }
-  let(:extractor_class) { ConnectorsSdk::SharePoint::Extractor }
-  let(:custom_client_error) { ConnectorsSdk::Office365::CustomClient::ClientError }
+# TODO: do proper mocking
+RSpec.describe ConnectorsSdk::SharePoint::HttpCallWrapper do
+  # XXX This is also stubs in lib/stubs/app_config.rb
+
   let(:backend) do
-    wrapper_class.new
+    described_class.new
   end
 
   let(:params) do
     {
-      :cursors => {},
       :access_token => 'something',
       :index_permissions => true
     }
   end
 
+  def mock_endpoint(path, data)
+    data = JSON.generate(data)
+
+    stub_request(:get, "https://graph.microsoft.com/v1.0/#{path}")
+      .with { true }
+      .to_return(status: 200, body: data)
+  end
+
   context '.document_batch' do
     it 'can get documents' do
-      allow_any_instance_of(extractor_class).to receive(:yield_document_changes)
-        .and_yield(:create_or_update, 1, nil)
-        .and_yield(:create_or_update, 2, nil)
-        .and_yield(:create_or_update, 3, nil)
-        .and_yield(:create_or_update, 4, nil)
-        .and_yield(:create_or_update, 5, nil)
+      # fake data
+      sites = { value: [{ id: 4567 }] }
+      groups = { value: [{ id: 1234, createdDateTime: Time.now }] }
+      drives = { value: [{ id: 4567, driveType: 'documentLibrary' }] }
+      drive = { id: 1111, driveType: 'documentLibrary' }
+      children = { value: [{ folder: 'folder', id: 1111, name: 'item' }] }
+      permissions = {
+        value: [
+          { id: 666 }
+        ]
+      }
+
+      mock_endpoint('sites/?$select=id&search=&top=10', sites)
+      mock_endpoint('groups/?$select=id,createdDateTime', groups)
+      mock_endpoint('groups/1234/sites/root?$select=id', sites)
+      mock_endpoint('sites/4567/drives/?$select=id,owner,name,driveType', drives)
+      # ??
+      mock_endpoint('sites//drives/?$select=id,owner,name,driveType', drives)
+      mock_endpoint('drives/4567/root?$select=id', drive)
+      mock_endpoint('drives/4567/items/1111/children', children)
+      mock_endpoint('drives/4567/items/1111/permissions', permissions)
+
+      extractor = backend.extractor(params)
       results, _cursors, _completed = backend.document_batch(params)
 
-      expect(results.size).to eq 5
+      expect(results.size).to eq 100
+      expect(extractor.config.index_permissions).to be_truthy
     end
   end
 
@@ -54,7 +83,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
     end
 
     before(:each) do
-      allow(extractor_class).to receive(:new).and_return(extractor_mock)
+      allow(ConnectorsSdk::SharePoint::Extractor).to receive(:new).and_return(extractor_mock)
     end
 
     it 'calls extractor.download method with same params' do
@@ -86,7 +115,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
     end
 
     before(:each) do
-      allow(extractor_class).to receive(:new).and_return(extractor_mock)
+      allow(ConnectorsSdk::SharePoint::Extractor).to receive(:new).and_return(extractor_mock)
     end
 
     context 'with valid access token' do
@@ -98,7 +127,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
 
     context 'with invalid access token' do
       it 'raise InvalidTokenError' do
-        allow(extractor_mock).to receive(:yield_deleted_ids).with(ids).and_raise(custom_client_error.new(401, nil))
+        allow(extractor_mock).to receive(:yield_deleted_ids).with(ids).and_raise(ConnectorsSdk::Office365::CustomClient::ClientError.new(401, nil))
         expect { backend.deleted(params) }.to raise_error(ConnectorsShared::InvalidTokenError)
       end
     end
@@ -115,7 +144,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
     end
 
     before(:each) do
-      allow(extractor_class).to receive(:new).and_return(extractor_mock)
+      allow(ConnectorsSdk::SharePoint::Extractor).to receive(:new).and_return(extractor_mock)
     end
 
     context 'with valid access token' do
@@ -128,7 +157,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
 
     context 'with invalid access token' do
       it 'raise InvalidTokenError' do
-        allow(extractor_mock).to receive(:yield_permissions).with(user_id).and_raise(custom_client_error.new(401, nil))
+        allow(extractor_mock).to receive(:yield_permissions).with(user_id).and_raise(ConnectorsSdk::Office365::CustomClient::ClientError.new(401, nil))
         expect { backend.permissions(params) }.to raise_error(ConnectorsShared::InvalidTokenError)
       end
     end
@@ -139,7 +168,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
 
     context 'remote source is up' do
       it 'returns OK status' do
-        allow(backend).to receive(:health_check).and_return({})
+        allow_any_instance_of(ConnectorsSdk::Office365::CustomClient).to receive(:me).and_return({})
         response = backend.source_status(params)
         expect(response[:status]).to eq 'OK'
       end
@@ -147,7 +176,7 @@ describe ConnectorsSdk::Base::HttpCallWrapper do
 
     context 'remote source is down' do
       it 'returns FAILURE status' do
-        allow(backend).to receive(:health_check).and_raise(StandardError)
+        allow_any_instance_of(ConnectorsSdk::Office365::CustomClient).to receive(:me).and_raise(StandardError)
         response = backend.source_status(params)
         expect(response[:status]).to eq 'FAILURE'
       end
